♻ refactor(api): ajustes nas camadas
